move getCourseName to course_cleaning
