+++ src/commands/owner/csClassPoll.ts
@@ -9,7 +9,7 @@ import { ICommand } from "wokcommands";
 import { classModel, IClass } from "../../models/classModel";
 import { checkForRoles, cleanRoleString } from "../../utils/roles";
 import { sleep } from "../../utils/sleep";
-import { getCourseName } from "../../utils/channels";
+import { getCourseName } from "../../utils/course_cleaning";
 
 // Splits any size list into lists of at most `max_list_len`
 /**
